requirements.txt
bcrypt==4.3.0
cffi==1.17.1
cryptography==41.0.7
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.1.1
